Add files via upload
Added parallel processing  and the removal of outlier samples  into the diff_expr  function. Also add pathway analysis via EnrichR on  top 20 genes. This can be performed on all 20, those that are upregulated , and those that downregulated . I also added a volcano plot and a UMAP.  Also all plots are interactive with hover functionality and related sliders. Further details will be shown in the new Readme.